+++ src/main/java/me/despical/oitc/commands/admin/arena/ForceStartCommand.java
@@ -1,6 +1,6 @@
 package me.despical.oitc.commands.admin.arena;
 
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import org.bukkit.command.CommandSender;
@@ -27,7 +27,7 @@ public class ForceStartCommand extends SubCommand {
 
 	@Override
 	public String getPossibleArguments() {
-		return "";
+		return null;
 	}
 
 	@Override
@@ -45,19 +45,18 @@ public class ForceStartCommand extends SubCommand {
 			getPlugin().getChatManager().broadcast(arena, getPlugin().getChatManager().formatMessage(arena, getPlugin().getChatManager().colorMessage("In-Game.Messages.Lobby-Messages.Waiting-For-Players"), arena.getMinimumPlayers()));
 			return;
 		}
+		
 		if (arena.getArenaState() == ArenaState.WAITING_FOR_PLAYERS || arena.getArenaState() == ArenaState.STARTING) {
 			arena.setArenaState(ArenaState.STARTING);
 			arena.setForceStart(true);
 			arena.setTimer(0);
-			for (Player p : ArenaRegistry.getArena((Player) sender).getPlayers()) {
-				p.sendMessage(getPlugin().getChatManager().getPrefix() + getPlugin().getChatManager().colorMessage("In-Game.Messages.Admin-Messages.Set-Starting-In-To-0"));
-			}
+			ArenaRegistry.getArena((Player) sender).getPlayers().forEach(p -> p.sendMessage(getPlugin().getChatManager().getPrefix() + getPlugin().getChatManager().colorMessage("In-Game.Messages.Admin-Messages.Set-Starting-In-To-0")));
 		}
 	}
 
 	@Override
 	public List<String> getTutorial() {
-		return Arrays.asList("Force start arena you're in");
+		return Collections.singletonList("Force start arena you're in");
 	}
 
 	@Override